🔨🤖 (bespoke) give causes-of-death a core/config.ts
It was the only project parsing its ArchieML config inline in mount. The
config interface moves out of CausesOfDeathConstants.ts, and year now parses
like the other projects (Number + finite check), so a malformed value falls
back to undefined instead of a truncated parseInt result.
