+++ bespoke/projects/causes-of-death/src/components/CausesOfDeathChart.tsx
@@ -11,7 +11,7 @@ import * as R from "remeda"
 import { Time } from "@ourworldindata/types"
 import { WORLD_ENTITY_NAME } from "@ourworldindata/grapher/src/core/GrapherConstants.js"
 
-import { CausesOfDeathConfig } from "../core/CausesOfDeathConstants.js"
+import { CausesOfDeathConfig } from "../core/config.js"
 import {
     useCausesOfDeathEntityData,
     useCausesOfDeathMetadata,
