+++ bespoke/projects/causes-of-death/src/core/CausesOfDeathConstants.ts
@@ -101,12 +101,3 @@ export interface TooltipState {
     target: { node: TreeNode } | null
     position: { x: number; y: number }
 }
-
-export interface CausesOfDeathConfig {
-    region?: string
-    sex?: string
-    ageGroup?: string
-    year?: number
-    hideControls?: boolean
-    urlSync?: boolean
-}
